refactor(tf2-format): reorganize schemas for better structure
