+++ libs/tf2-format/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/types';
 export * from './lib/parsing';
 export * from './lib/sku';
 export * from './lib/naming';
+export * from './lib/schemas';
